requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.1
plotly>=5.15.0
numpy>=1.24.0
scipy>=1.10.0
seaborn>=0.12.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
python-dateutil>=2.8.0
pytz>=2023.3
groq>=0.4.0
python-dotenv>=1.0.0 